Refactor: ScrollTopButton Component 분리
- ScrollTopButton Component 생성
- MainPage/components index.js 생성
