+++ src/pages/MainPage/components/index.js
@@ -1,0 +1,3 @@
+export { default as MainGrid } from "./MainGrid";
+export { default as MainInput } from "./MainInput";
+export { default as ImageData } from "./SliderImage/ImageData";
