+++ src/pages/PostListPage/components/ScrollTopButton/ScrollTopButton.style.js
@@ -1,0 +1,10 @@
+import styled from "@emotion/styled";
+import { Button } from "components";
+
+const ButtonWrapper = styled(Button)`
+  position: fixed;
+  right: 10px;
+  bottom: 10px;
+`;
+
+export default { ButtonWrapper };
