+++ src/pages/PostListPage/components/ScrollTopButton/index.js
@@ -1,0 +1,3 @@
+import ScrollTopButton from "./ScrollTopButton";
+
+export default ScrollTopButton;
